backend/pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = [
    {name = "Didenko Victoria",email = "155799784+misfits3z@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "sphinx (>=8.2.3,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

